requirements.txt
numpy
pillow
opencv-python
pypotrace
flask

notebook
pre-commit # for formatting